Implement the appropriate LoopProgress progress bar. Refactor the error handling logic in the code.
