+++ vulnerabilities/tests/pipelines/test_exploitdb.py
@@ -7,7 +7,7 @@ import pytest
 from vulnerabilities.models import Alias
 from vulnerabilities.models import Exploit
 from vulnerabilities.models import Vulnerability
-from vulnerabilities.pipelines.exploitdb import ExploitDBImproverPipeline
+from vulnerabilities.pipelines.enhance_with_exploitdb import ExploitDBImproverPipeline
 
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 TEST_DATA = os.path.join(BASE_DIR, "../test_data", "exploitdb_improver/files_exploits.csv")
